refactor(mpsc): keep errors local
